backend/requirements.txt
# Django et Django REST Framework
Django==4.2.7,<5.0
djangorestframework==3.14.0

python-decouple==3.8
django-jazzmin==2.6.0
Pillow==10.0.0
django-filter==23.3

# PostgreSQL
psycopg2-binary==2.9.9

# Gestion des environnements et configurations
# python-dotenv==1.0.0
# dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.0

# Sécurité
# django-environ==0.10.0

# Performance et utilitaires
# gunicorn==21.2.0
# whitenoise==6.6.0